1658. Minimum Operations to Reduce X to Zero
